Calculations for CfD simulations
The TradeRES_bbresutls.py script includes the BackboneResult and BackboneScenarioAnalysis classes written by David Huckebrink that are used to convert Backbone outputs (gdx-files) to Pandas DataFrames. I added some plot functions to the classes that I used in the TradeRES project.
The TradeRES_CfDs.ipynb jupyter notebook was used to Backbone input excels with updated costs and capaciy limits for simulations of different types of CfDs.